requirements.txt
aerich==0.4.2
aiosqlite==0.16.0
asyncpg==0.21.0
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.62.0
fastapi-users==3.1.1
h11==0.11.0
idna==2.10
iso8601==0.1.13
itsdangerous==1.1.0
makefun==1.9.5
passlib==1.7.4
pycparser==2.20
pydantic==1.7.3
PyJWT==1.7.1
PyPika==0.44.1
python-multipart==0.0.5
pytz==2020.4
six==1.15.0
starlette==0.13.6
tortoise-orm==0.16.18
typing-extensions==3.7.4.3
uvicorn==0.13.1
